requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
langchain==0.2.11
langchain-community==0.2.10
faiss-cpu==1.7.4
ollama==0.1.7
unstructured==0.10.28
pdfplumber==0.11.0
pypdf==3.12.2
crewai==0.28.8
google-api-python-client==2.137.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
openai>=1.42.0
langchain_groq
langchain_ollama
groq